Abstracted Movie Main Info in two differents widgets. One abstraction for the movie image and other abstraction for the movie info.
