+++ example/lib/src/resources/route_constants.dart
@@ -1,4 +1,5 @@
 const String homeRoute = '/';
 const String customAppBar = '/customAppBar';
 const String movieDescription = '/movieDescription';
+const String movieDetailsImage = '/movieDetailsImage';
 const String movieMainInfo = '/movieMainInfo';
