+++ example/lib/src/utils/ui_constants.dart
@@ -4,7 +4,8 @@ class UiConstants {
       'https://www.themoviedb.org/t/p/w1280/qa6HCwP4Z15l3hpsASz3auugEW6.jpg';
   static const String customAppBarTitle = 'Custom App Bar';
   static const String movieDescriptionTitle = 'Movie Description';
-  static const String movieMainInfoTitle = 'Movie Main Info';
+  static const String movieMainInfoTitle = 'Movie Details Info';
+  static const String movieDetailsImageTitle = 'Movie Details Image';
   static const String defaultOverView =
       'Lightning McQueen, a hotshot rookie race car driven to succeed, discovers that life is about the journey, not the finish line, when he finds himself unexpectedly detoured in the sleepy Route 66 town of Radiator Springs. '
       'On route across the country to the big Piston Cup Championship in California to compete against two seasoned pros, McQueen gets to know the towns offbeat characters.';
@@ -16,7 +17,7 @@ class UiConstants {
 
   //home_screen
   static const String appBarTitle = 'Widgets Mini App';
-  static const double containerTopPadding = 55.0;
-  static const double containerHeight = 130.0;
+  static const double containerTopPadding = 40.0;
+  static const double containerHeight = 100.0;
   static const double containerTitleFontSize = 40.0;
 }
