+++ lib/movies_project_widgets.dart
@@ -2,5 +2,5 @@ library movies_widgets;
 
 export 'src/widgets/custom_appbar.dart';
 export 'src/widgets/movie_description.dart';
-export 'src/widgets/movie_details_page.dart';
+export 'src/widgets/movie_details_image.dart';
 export 'src/widgets/movie_main_info.dart';
