+++ lib/src/utils/ui_constants.dart
@@ -5,9 +5,6 @@ abstract class UiConstants {
   static const double movieTypeTitleFontSize = 25.0;
   static const double progressIndicatorStrokeWidth = 5.0;
 
-  // widgets/movie_details_page
-  static const double detailsSizeBoxHeight = 10.0;
-
   // widgets/custom_appbar
   static const double movieDetailsAppBarExpandedHeight = 350.0;
 
@@ -16,11 +13,12 @@ abstract class UiConstants {
   static const double descriptionFontSize = 18.0;
 
   // widgets/movie_main_info
-  static const double widgetPadding = 20.0;
-  static const double posterPathBorderRadius = 20.0;
-  static const double posterPathHeight = 150.0;
   static const double movieInfoSizeBox = 10.0;
   static const double movieTitleFontSize = 30.0;
   static const double releaseDateFontSize = 20.0;
   static const double voteAverageFontSize = 36.0;
+
+  // widgets/movie_details_image
+  static const double posterPathHeight = 150.0;
+  static const double posterPathBorderRadius = 20.0;
 }
